moved the minutes title to its own component
